Refactor HMM according to PairHMM
Disconnect HMM class from the hierarchy. Add the interface as in
PairHMM. Refactor algorithms to comply with PairHMM implementation
so it will be easier to join the classes later.

Refactor HMM tests and change helpers to generate a trained and
untrained coin dishonest casino example.

Change the definition of 'Label' to be a struct so that it is possible
to unpack it within for-ranged loops.